backend/requirements.txt

Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
pillow==10.2.0
numpy==1.26.4
tensorflow==2.15.0
python-magic==0.4.27
